remove lab/ -- create the generated modules with make
Seems like the best thing is to get rid of the lab/ modules 
from the repo for now. They can be generated ny make.
We can certainly find a way to host them reliably close 
by or in another branch --but for now they are causing 
more trouble then they are worth.