workaround for canvas/zIndex bug ?? in Chrome Canary
Before this change black borders appear on the top of 
the KE graph, Distribution of Speeds, and Molecular container
divs when change the number of molecules. The artifacts appear 
when setting the zIndex on the canvas overlay used in 
the KE graph to -100.

This change in zIndex is performed whenever the model is
paused -- but unless the molecule number is changed the
artifacts do NOT appear when pausing the model.

If the page is loaded AND the model paused … then the artifacts
appear when performing just a mousedown on molecule number 
selector -- this is BEFORE the molecule number is changed
(the onchange event hasn't fired yet).

This bug is present on Chrome Canary 18.0.10254.1 on 
Mac OS X 10.6.8. Not present on Chrome 16.0.912.21 beta,
Safari 5.1.2 (6534.52.7), or FireFox 9.0.1

To reproduce generate examples with ANY commit before 
this one and open either:

  examples/complex-atoms-model/complex-atoms-model.html

Or if using older commits before this example was present:

  examples/simplemolecules/simplemolecules.html

Open page, press spacebar to stop model, mousedown on
molecule number selector.

screenshot: https://img.skitch.com/20120201-jqa77etmnsxkupeywx4896qny1.jpg

This is a very confusing bug, it's not at all clear why 
this change  fixes the problem, was unsuccessful trying
to create a simplified test case.