Remove generate_atoms from modeler.js
Access the node array of atom property arrays more directly
from the view. The 'atoms' array is now just used as an array
with the length of the number of atoms being simulated.

By not recreating the atoms array with individual atom objects
for each tick we should reduce the number of objects that need
to be garbage collected.