modify document.domain after loading rest of code
This normally works fine but is now giving a security error in FireFox:

  >>> document.styleSheets[3]
  StyleSheet examples.css

  >>> document.styleSheets[3].cssRules
  [Exception... "Security error" code: "1000" nsresult: "0x805303e8 (NS_ERROR_DOM_SECURITY_ERR)" location: "<unknown>"] { constructor=DOMException, code=1000, INDEX_SIZE_ERR=1, more...}

Turns out that FF does this when document.domain is modified -- see the last comment in this FF issue from 2008:

  https://bugzilla.mozilla.org/show_bug.cgi?id=470321

I had modified document.domain so we could create demos which 
communicate through the iframe on concord.org domains.

The security problem does not happen locally. So I'll commit
and deploy this to test on the external server.