Simple Atoms model now use simpleModelController
THe application.js code uses by the three Simple Atom model
variants has shrunk to almost nothing. The code was refactored
into a primitive controller: simpleModelController.

In addition the molecular container is converted into a closure-pattern
so we are much closer to being able to create multiple molecular
containers.

The main layout system got simpler.

Lastly the SliderComponent is fixed so that the min, max, and initial values
can be passed on initialization. In addition the position of the handle on the
slider is now correctly placed -- previously it was always in the center.