[#27225477] use closure pattern for complexModelController
Now creating a complexModelController returns an object
so patterns like this are both possible and required
(fewer global variables):

model_controller = controllers.complexModelController("full-static-screen");

Now the model_controller object is what is listening to the view
and communicating changes to the model. 

The model_controller also implements the modelListener
which is called by the simulation after completion of
a model step and asks the view objects to update themselves.

This also means that if some other code needs to run methods
in the controller the code needs to use controller methods
exposed on the controller object.

For example the benchmarking code uses the model_controller
object to call the modeStop() command.

  model_controller.modelStop();