Start of work to allow serialization and loading of properties
This is the beginning of a pattern where all the basic properties
of a model (thermostat temperature, existence of coulomb forces, etc.)
are contained privately within the model object, and are accessible
through a standardized setter and getter.

The setter can accept a hash of properties, allowing us to load in
an entire json object from a saved model.

The serialize function on model allows us to save all the properties
in a stringified json object, which can be loaded back.