Fix function for setting value in slider.coffee
Note: there seems to be some confusion in the code about the
value and the scaled_value. When the view is first initialized,
it is passed -0.1 and this is set as the value, but based on
what happens after the slider is dragged, it seems that this
was meant to be the scaled_value.