Update Makefile to convert MML files and make html page
This is probably not set up the way we will want it in
the future. When 'make clean; make' is run, the converter
will run and add the converted files to the imports/ dir,
then the html page will be made in the same dir, and then
the whole dir will be copied over to dist.

Note the way it is set up now if 'make clean' is not run
first, the conversion script won't ever run, I think.