Change the LJ calculator to use pairwise epsilons and sigmas between elements
All parameters in the LJ calculator, including things such as rmin and the
cutoff-distance are now only defined as matrices defining the values between
any pairs of elements.

This breaks the notion of having any kind of model-level epsilon and sigma. Any
functionality that depended on this (pretty much just the LJ slider/graph) will
not work until they start trying to affect the e's and s's at the element level.

The benchmark test showed no significant slowdown after this change, but running
the complex model appears to be slower, at least at the start. More optimization is
required. (Low hanging fruit include changing the element arrays to be typed.)