Guardfile/Makefile: make src, make test-src
The build process after a change in the src/ dir
had become unreasonably long because of the mw-import
script running anytime a change was made.

These changes greatly reduce the time the browser can
be reloaded after making a change.

There will probably need to be more work ... there may
be changes that are made in some sections of the source
that don't automatically kickoff the right actions.