Many kinds of data can be loaded into Graphing Example
I've made the Simple Graph page into a Graphing Examples
page. It now supports the data displayed in the original
Sample Graph as well as datasets from the Earth's Surface
Temperature and a Center of Mass Analysis from the MD2D
work.

In addition there is a simple streaming data example.