use .gitattributes for per-file merge strategy
When merging from one branch to another always resolve
any possible merge conflict in .travis-yml in favor of
the branch being merged into.

This allows different notification lists to be maintained
for travis-ci build notifications per branch.

However this also means more manual effort when there is
a change in .travis.yml that should be propagated to
all branches that integrate with travis-ci.