gpgpu experiment: make sure alpha channel is enabled
lightgl library disables alpha channel as default

In this gpgpu  test on Chome v20 on Mac lack of alpha setting
caused image result of running encodeShader() to have
all alpha-channel bytes to be set to 255.

This resulted in NaN values when read back as Floats.

The problem did not occur on Windows Chrome 20, Ubuntu Chrome 21,
Mac Firefox 13, or Mac Chrome 22