Add ability to create spring force between an atom and a point
A spring force can be created between any atom and any x,y
location, with a spring constant k. After this, the x,y location
can be updated on the fly, and the spring force can be removed.

Note we will probably want to refactor/rethink how we can
update and remove individual members of arrays, as this
might be something we'd want to do for all arrays (atoms,
radialBonds, etc.).