add a "Start Moving" button to one-radial-bond
The "Start Moving" button adds a 0.00002 increase in 
X velocity to both atoms so the bonded pair start moving
slowly to the right.

This helps show two different issues:

1. When the bonded pair hit the wall they bounce away with
an incredible increase in velocity.

2. That the frequency of the atom pair velocity due to the
radial bond is NOT what it appears to be in the visualization.

However these are not exactly bugs ...