restore and fix model.reset()
[#35438399]

The first model state stored in the tic_history_list is
now saved and not overwritten. When a model.reset() happens
the tic_history_list is reset to length 1 with just the
first model state saved. This first model state is restored.

When the max buffer size for the tic_history_list is exceeded
the second saved state is deleted now, not the first.
So the first state is always preserved.