distinguish event and property listening
When dispatching *events* like seek or reset, use model.on(...) (which internally uses d3.dispatch) to notify listeners. notifyListeners... is only for listening to property changes.

We were attempting to use both pathways to notify of seek and reset events, but I believe this is a mistaken overloading of the property listener functionality. Possibly there is a way to extend model.on to listen for property changes as well.