generate new atom property MASS from element data
This exposes the new atom property in an array named 'mass'
so in physics calculations the mass of an atom can be referenced:

  mass[atom_index]

This value for this mass is set based on the atoms
element type and the mass for that element type.

The model runs approximately 2-5% faster.

The code where mass is used is a bit easier to read.