create transposed output arrays for Molecular view
Transposed arrays named results and radialBondResults are
updated when computing MD2D outputState and made available
to the Molecular view. 

The 'results' arrays consists of arrays of individual atom
properties and starts with an atom index value.

A similar array named radialBondResults is created that includes
an index value for the specific radial bond, the standard
radial bond attributes as well as the positions of the two
bonded atoms.

The MD2D engine still uses the individual property arrays for
all the physics calculations. The new 'results' arrays are
created during the outputState phase and are structured in a
way so the D3-based view can be more efficient.

So far there appears to be about a 5% improvement in the
combined model and graphics benchmark.