add extract of blog post to Interactive Sampler
[#36786727]

There is an iframe layout issue in Chrome 21.0.1180.75
that is not present in Safari 5.1.7 or Firefox 15.01.

When the page with this sample is first loaded in Chrome
the scroll bars are still visible as if the iframe has
not been re-sized.

Opening webkit inspector and looking at the iframe content
causes the resize event to be propogated and the layout corrected.