use actual model aspect ratio in main Interactive layout
[#37159927]

The layout system for the non-embedded Interactives now
gets the actual model aspect ratio. 

The Thermometer height no longer needs to be actively changed
because emsize is set on the #content div the main Interactive
is embedded into.

The energyGraph and viz css modifications are commented out
because they don't appear to be needed anymore.

In addition the html structure of the main Interactive page
is changed so that only the content and components that will
appear in an embedded Interactive are in the #content div.