Layout should get dimensions from actual DOM element!
One issue: who knows for sure what methods the molecule container will implement in the future? We do know we can get its *actual* width and height using jQuery (Although to be fair we are expecting it to continue to export an 'outerNode' property that we can inspect using jQuery.)

Another issue: in fit-to-parent mode (used by standard layout) the first two elements of the array returned by scale() aren't measured in actual on-screen pixels. Rather we use a notional width (cx = 500) and allow the browser to scale that width up or down in order to fit into a percentage-width sized container.