integrate Radial Bond style persistence with MD2D
[#37595697]

This does NOT add any new styling. It just exposes an explicit coding
for the appropriate style in the radialBondsResults array.

The only change in the view is that this explicit coding is now use
to determine if the bond style should be a spring instead of the
previous heuristic algorithm.