new required method: model.initiaizeHistory();
Initialization of tick history moved to it's own method.

When creating an MD2D model call this method after
completing the initialization of the model and before
using it.

This will make sure that the tick history is setup
properly. A subsequent commit will add obstacles to
the tick history and without this re-factoring the first
state saved in the tick history was invalid.