Improve console implementation
Now, every component / class which uses console, should require it:
  var console = require('common/console');
and use provided console API (identical to the standard one).

It's wrapper around standard window.console implementation with an ability to enable or disable:
- logging (console.log, .info, .warn, .error)
- tracing (console.time, .timeEnd)

Methods which allow to do that are exported to the global namespace:
- Lab.console.logging
- Lab.console.tracing

For example, to change behavior dynamically, open the Web browser console and type:
  Lab.console.logging(true);
  Lab.console.tracing(true);
It will enable both logging and tracing.

When these functions are called without arguments, they will return current state of logging / tracing configuration.