VDWLinesCutoff: "short", "medium", "long"
[#38943097]

In order to speed up the view code in models which display
VDW attraction lines and have a large number of atoms in
close proximity add support for short, medium, and long
variable VDW Line Cutoffs.

This cuts the initial number of rendered VDW lines in
interactives/sam/phase-change-page-6.json from 780 to 408
which speeds up the Interactive by 50%.

The test VDW Conversion Test Interactive has been updated
to also show how VDWLineCutoff can be set with a script.

  interactives/conversion-tests/vdw-force-lines.json

TODO: the rendering is still incredibly inefficient becuase
for every view update it deletes and re-creates all the SVG
elements for the VDW lines.