Add experiment into defining component layout with html.
An author can now define a layout section (such as bottom) using
a string of html. The html will be embedded in the section, and
components will be added to this block of html by looking for dom
elements with appropriate ids.

By positioning these elements with %'s (or possibly by ems) an
author can position components such that they scale appropriately
with the model.

NB: If we adopt this we will probably want to change many of the
details -- this is being added as a quick spike to further discussion.
For example, components could be positioned using handlebars instead
of simply defining wrapping elements.

[#39286603]