Support categorization and filtering of Interactives
[#38171643]

NOTE: many Interactives have new simplified filenames and
some have been moved into new directories.

A new ruby helper: process-interactives.rb

  src/helpers/examples/interactives/process-interactives.rb

introspects into all the interactive json files and also
uses data in a new: groups.yaml file:

  src/examples/interactives/interactives/groups.yaml

to generate the following interactives.json file which is used
to automatically populate the Interactive select list.

  server/public/examples/interactives/interactives.json

This file can also be read by an external application to
discover the available Interactive resources. These data can
be used to dynamically generate appropriate categories of
Interactives for embedding into external applications.

  http://localhost:3000/examples/interactives/interactives.json

Currently groups.yaml file is used to categories Interactives
in specific directories as one of these types:

  Samples
  Curriculum
  Examples
  Tests

The categories: Curriculum, Examples, and Tests can be filtered
into or out of the select list using the new checkbox controls.

The order of array elements in the groups.yaml file control the
order of select option groups in the select menu.

In addition each interactive now has a publicationStatus field.

Valid values for publicationStatus:

  "sample"    these are ALWAYS included in the select list
  "public"
  "draft"

Interactives with publicationStatus categories of "public" or "draft"
can be filtered into or out of the select list using the new
checkbox controls also.

A persistent session cookie is saved which maintains the filter options
for the select list for a user navigating from one Interactive to
another on the same browser.

A second new field named "subtitle" has been added to every Interactive:

If the "subtitle" OR "about" fields of the Interactive have content
then an additional "?" link appears in the upper-right of the
Interactive which will open an AboutPane window with this content.

The default text previously used in the about field has been removed.
In order to enable the appearance of the new "?" link content must
be added to the subtitle or about fields.

Content in the about field can either be a single string OR and array
of strings. In the case of an array of strings each array element will
be rendered as a separate paragraph.

The "subtitle" and "about" fields for the default Oil and Water
Interactive have been filled in to provide a working example.

This Interactive has a new location in the samples directory:

  interactives/samples/1-oil-and-water-shake.json