even simpler instructionBox3.svg with drop shadow
[#40314483]

Another experiment ... this is hand-coded, simple, and less
than 1200 bytes. It uses museo-san for the font which is web-safe.
A drop-shadow is implemented with an SVG filter.

To test, edit the imageMapping for instructionBox.png in
the Interactive JSON.

For some reason it isn't scaling like the other examples,
and doesn't appear to work on FF.