content in Credits, Share, and About panes is selectable
[#40409363]

Limit the mouse-down dragging area in these Panes to the
banner-heading area in these dialog panes and use a CSS "move"
cursor to indicate when the cursor is over the draggable area.

The result of this change makes the content below the banner
selectable so it can be copied into the clipboard.