added enough About content to test scrolling
[#40409693]

The new Interactive for testing replacing png bitmap images
with SVG now has additional About content describing how
it works and what it is testing,

This content is long enough to require scrolling in the About
dialog.

In addition this About content is an example of how HTML can
be used effective in the About content.