set new elements to 0 when extending non-typed arrays
[#40792663]

When extending a typed-array we are setting new elements to 0.
Creating a new Obstacle by extending the Obstacle arrays
assumed this behavior for initializing properties such as
obstacles.westProbeValue.

However this was **not** the default behavior when extending the
length of a regular JavaScript array. This resulted in initial
values being undefined ... and causing errors of this type:

  Error: FilteredOuput input property is not defined or it is not valid numeric value.

when executing this code:

  initialValue = model.get(property);

which was getting the value of the westProbeValue property
from this array of obstacle property values:

  obstacles.westProbeValue