Support special forces affecting amino acids in oil or water env
[#40697971]

Note that it's initial implementation for tests. Many parts connected with proteins engine needs refactoring. I defer it to avoid premature refactoring, focus on physic and later cleanup code when we know exactly the final shape of this algorithm.

E.g. structure for storing molecules would be useful, as current implementation causes the calculation of molecule atoms really expensive and do not cache this information.