new embeddable-author Interactive page type
[#41395477]

There is now a new embeddable-author.html Interactive page
type to go along with the existing interactives.html and
embeddable.html.

An embeddable-author page includes the Interactive editor on
the left and the embedded Interactive on the right.

The page is designed to fit into the 938px x 400px
excercise area on a Khan CS page like this:

http://www.khanacademy.org/cs/intro-to-animation/830742281

Here's an example of an iframe using this new page type:

<iframe width="938px" height="400px" frameborder="no" src="http://lab.dev.concord.org/examples/interactives/embeddable-author.html#interactives/sam/gas-laws/3-volume-pressure-relationship.json"></iframe>