Visualize disulphide bonds in a special way (yellow color)
[#41092127]

Note that this work also includes a small refactoring of radial bond
styles and a new SASS file. This file is intended to contain CSS
styles of elements managed by molecule-container, which can be defined
statically, without JS.

Probably it can be a good idea to move there as much as it's possible
to make visual tweaking of the simulation faster and easier.