benchmarks: generate averaged values in 2nd row
[#42103253]

After running the performance benchmarks the second row in the
table contains the averaged values of all sample runs. The averaged
data row is styled with italic text and a light green background.

Two additional properties have been added to the benchmark definition
object:

- numeric
A boolean indicating whether the value returned by this
test is a number.

- formatter
If present this should be a function that takes a numeric
input and returns a formatted string.