Guardfile/Makefile: new ENV var: LAB_DEVELOPMENT
[#42157809]

When the LAB_DEVELOPMENT ENV var is set the Makefile
skips running JavaScript minimization.

THe Guardfile exports LAB_DEVELOPMENT when active.

In addition the Makefile target test-src (which the Guardfile runs
when changes are made in src/lab or src/modules) now skips running
the slow Mocha tests.

All tests are run when running: make test and any individual
Mocha test will be run automatically by the Guardfile when a
change is made in the specific test.

If you are running make targets outside of Guard you can also pass
the variable LAB_DEVELOPMENT as an argument to the make target invocation:

Examples:

  make LAB_DEVELOPMENT=true

  make public LAB_DEVELOPMENT=true

Erase and re-generate the server/public folder