example of SVG doc using Open Sans font
[#42229129]

/resources/pressureGauge.svg

When using this same form as an attached image the font
specification doesn't appear to be resolved properly.

Test details: copy /resources/pressureGauge.svg to:

/imports/legacy-mw-content/sam-activities/gas-laws/original-interactives-in-pages/page5/pressureGauge.svg

and rebuild the project.

This file is referenced from this Interactive:

/examples/interactives/interactives.html#interactives/sam/gas-laws/5-temperature-pressure-relationship.json

However when rendered now in the Interactive pressureGauge.svg still
uses a serif font.