don't need Z-index, make sure row doesn't cover logo
[#42435089]

By making the display property for #header.row 'inline block'
the row stops at the end of it's content and doesn't
start to cover up the logo link.

Before this fix the logo link was clickable only at the very top
where it wasn't covered by the row of additional links.