Wrap text to given width if specified in textBox.
Add helper function to wrap text such that each line
at least one word long but (if possible) is less or
equal to the desired length when rendered. Note that
getComputedTextLength() will return a pixel value, but
this value depends on the current size of em, and so
scales correctly with our view.

Note that textBox widths do not exist in MML, so we
do not convert any, and so currently the only way to
see this is to edit the modelOptions in the interactive.

[#37610629]