For now, completely replace arrays in interactive's options.
If we are replacing properties of the original model
using the viewOptions or modelOptions in the interactive,
specifying an array property will not completely replace
the original array, instead of only replacing the first
n elements of that array.

This seems to follow the principle of least-surprise for
what an author would expect.

If needed, we could add a function in the future for an
author to be able to replace or amend specific existing
elements of the original array.