add setFocus to MD2D view, use on non-embed page load
[#42517011]

Now when using the Next/Previous buttons to sequence through
Interactives after the Interactive loads the keyboard
focus is set to the MD2D view.

This allows a user/tester to start or stop the model using
the keyboard controls.

The MD2D view option: enableKeyboardHandlers must be true
for the keyboard controls to work. This is set true by default
only when the play_reset_step playback control is used.