Serialize genetic properties, change metamodel of g. properties
[#43089109]

Also, now translationStep and mRNA are immutable values. This means
that they cannot be modified directly by setGeneticProperties()
function, so user will be forced to use provided API (e.g.
transcribe() or translateStepByStep()).

Previously these properties were readOnly, but that caused that
serialization and correct loading of state was impossible.