Downsample force sensor
We aren't able to pass multiple data points into the current dataset without also emitting multiple 'data' events. Thus we are forced to redraw the graph 5 times for a single sensor callback that contains 5 data points, or else to downsample the data.

Moreover downsampling is itself made a little awkward by the dataset class because we are forced to set x increments for the raw and the filtered data, even though we ought to be able to just grab x-values from the raw data stream.