Fix ParentMessageController initialization
Previous way of initialization was causing that each time initialize()
was called, a new listener of "message" event was added.
So, single message was triggering callback many times.

ParentMessageController.initialize() is called for example during
.loadInteractive or  .loadModel methods of InteractiveController.
So clicking multiple times "Load Interactive" was causing a lot of
problems and sometimes crashing the browser.

TODO: should we ParentMessageController.initialize() each time a model
is loaded?