Slightly change layout algorithm in semantic layout experiment
[#44281603]

Resizing of the model container + repositioning is in fact solving of
a linear equation needed to provide correct layout. Because of that,
resize model in a different way, which provides better convergence.
(TODO: still it is far from perfect, it's just experiment)

Previous method worked for most cases very precisely, however caused
problems for more complex setups, e.g.:

 {
  "id": "wide-right",
  "left": "model.right",
  "height": "model.height",
  "width": "model.width", // !!!
  "components": [
   "button2",
   "button3"
  ]
 }