restore dynamic actualRoot functionality
IS [#44317763]

When static resources in the examples, experiments, or
doc directories are delivered by a web server *and* those
resources are *not* in the root directory of the server …

Then lab.config.actualRoot will contain the path to 
the sub-dir where Lab resources are being hosted.

This should now also restore support for situations where
an external web site has setup a proxy at a sub-dir
pointing to a Lab server (such as DataGames).