add pulldown that uses property/value option
[#44812673]

/examples/interactives/interactives.html#interactives/conversion-tests/gravitational-field-test-property-pulldown.json

Interesting problem getting this to work …

As imported from Java MW the default gravitational field for
the model was: 8.000000000000001e-7

The values in the pulldown assume 8e-7.

I needed to add an "onLoad" option in the models section
to get the pulldown to work.

Without this change the pulldown displayed the "Gravitational Field Off"
option which is associated to the value: false.