use same initial vy for obstacle in gravity test
IS [#44983919]

All atoms and objects (that move) should have same initial vy.

The simple test for gravity working is that everything 
rebounds back to the same height and at the same rate
even though the atoms have different masses and
and moving obstacle is included also.

There are now two similar test Interactives for gravitational fields:

- /examples/interactives/interactives.html#interactives/conversion-tests/gravitational-field-test.json
- /interactives.html#interactives/conversion-tests/gravitational-field-test-property-pulldown.json

This exposed a physics problem:

When the gravitational field is set to small values,
for example: 2e-8, which is labeled as "Gravitational Field: 5"
in the pulldown menu and the model started the particles bounce
back higher than they started.