Call modelLoaded after modelController variable is added to the API
[#45030209]

It was causing errors in application.js, which was expecting that
modelController property of a interactive controller was available
when modelLoaded callback was executed.

In general, IMHO that also presents an anti-pattern of doubling
public variables and exporting them using:
api.someVariable = someVariable;
where someVariable is a closure variable used elsewhere.