Prevent model.start() from continuously speeding up model
Each time model.start() is called (say from an author
script) a new tick timer is added. This increases the speed
of the model.

This can be seen very clearly if you create an author
script that simply calls start() a bunch of times.

This fixes this behavior by returning immediately if
the model is already running.

[#45110023]