Optimize last slow mocha test
[#45143279]

This time we can't simply use withIsolatedRequireJSAndViewsMocked, as
we need layout module to attach view elements to the document.
However, we do not care about the layout quality, so we can require
and modify the layout configuration using layout config module. It's
enough to limit maximum number of allowed layout algorithm iterations
to speedup test significantly.