Trigger dragend when mouse leaves interactive container.
We check the target of the drag event and see whether
it is either an svg element or a dom element contained
within the #interactive-container. If it isn't, fire
a mouseup event to trigger d3's dragend event.

This has been tested in Chrome, FF and Safari.

Note this hard-codes the selector "#interactive-container".
This same selector is being hard-coded in the semantic-
layout (which doesn't appear to use the selector passed in
by application.js to the interactive controller).
TODO: Use the selector passed in by application.js and
make it a variable that's accessible in other parts of
the code.