Prevent second liveDrag force from being added.
Previously, if we missed the mouse-up command and the
atom was stuck to the mouse, a second click would add
a second spring force. Only one of these forces would
be removed on subsequent mouse-up, and the atom would
be effectively pinned to the point of the last spring.

This fixes that by preventing the addition of a second
liveDrag spring force when one is already in progress.
This does not prevent the model from adding other
concurrent spring forces.

[#45233971]