Atom dragging: disable display of user selection
[#45233971]

One of the unfortunate results of the work to get atom
dragging working when extending outside of an iframe
boundary is that mousedown events have need to bubble
up to the window level.

This causes an unfortunate effect of making selectable text
and elements highlight to show their selection.

This will need to be adapted when and if we need to support
selection in the model container -- for example entering
text for a label. This is likely to come up when we start
work on WYSIWYG model editing.