Add cachebusting to images to resolve Chrome 26+ bug
Chrome 26 and above has a bug where the image size
is cached after the size is changed in the code.

This can be seen at http://jsfiddle.net/SamFent/MmSrn/ (no bug report raised yet).

This resolves the issue by using a different image
url each time we download an image in Chrome 26+

[#45397937]