initial integration of model dimensions
[#41923323]

Support model dimensions expressed as:

  minX, minY, maxX, maxY

Support models and rendering systems where the lower-left
corner of the model is *not* assumed to always be 0,0.

Currently in MD2D the lower-left corner of the model
is assumed to be 0,0 and the minX, minY, maxX, maxY
variables are derived from the model parameters for
width and height.

There are now four model-to-pixel scaling functions,
two for position and another two for size:

model2px
  Basic scaling functions for position, it transforms model units to "pixels".
  Use it for positions of objects rendered inside the view.

model2pxInv
  Inverted scaling function for position transforming model units to "pixels".
  Use it for Y coordinates, as Y axis in model coordinate system increases
  from bottom to top, while but SVG has increases from top to bottom

modelSize2px
  Basic scaling function for size, it transforms model units to "pixels".
  Use it for dimensions of objects rendered inside the view.

modelSize2pxInv
  Inverted scaling function for size transforming model units to "pixels".
  Use it for Y coordinates, as Y axis in model coordinate system increases
  from bottom to top, while but SVG has increases from top to bottom