realTimeGrapher works better w/semantic layout
[#45855867]

The fonts for graph title, axis labels, and axis numbers
are now set in new sass file included into the main
lab sass file.

By default fonts are now displayed at whatever size is specified
in the css styles combined with the basic font size in the
container the graph is placed in.

In addition the responsive layout feature where certain
elements are deleted as the size of the graph becomes
smaller is off by default. There is a new 'responsiveLayout'
option to enable this feature.

Lab.grapher.realTimeGraph, new options:

- responsiveLayout: false,
- fontScaleRelativeToParent: true,