Support image urls starting with "/resources"
Previously image urls were either absolute (http://..)
or were relative to the model url.

This change allows urls of the form "/resources" to be
used without changes (i.e. making them relative to the
lab url) because it will be common to refer to images
in the main resources folder.