Pendulum and Spring: improve 2-graph layout
[#45966989]

With the limited space available when showing 2 graphs
on the right try a layout which removes the titles
and also the X axis labels.

The titles basically appear in the Y axis labels and
the numbers are still available on the X axis and the
unit "s" is visible in the model.