Add ability to pass in new container for playback controls
We can pass the model controller a new container for
playback controls. If we do, the view will no longer
render to its own bottom bar but instead will render
to this new container.

When the model view renders the controls to an external 
container, it now takes up all the available space.

If we do not pass in a container, the view will render
its controls as usual. This way we can continue to use
this feature when we starting having multiple models 
on the screen.

[#44616109]