Add fullscreen button to bottom bar.
Fullscreen button checks to see if the browser supports
the html5 fullscreen commands (not hacking with ActiveX
with IE here), and, if so, adds button div to bottom
bar of interactives.

iframe has to use the allowfullscreen option for this
to work. Update our default iframe and the sharing
text.

Currently we can't tell if iframe has this option
selected, so if iframe is not embedded correctly, button
will show but will do nothing.

[#39555821]