netlogo applets in Windows can't use paths
NetLogo applets running in Windows incorrectly convert 
the "/" in the path part of a url to "\".

So add a hack to the Makefile to copy data-export-modular.nls
to the same directories the main NetLogo models are located.

Also tell git to ignore the resulting copied files.

Example:

This is the url for an http request sent by a NetLogo applet
running in Windows for a resource.

http://visual.concord.org/experiments/netlogo-visual/models/..\..\netlogo-modules\data-export-modular.nls